remove the `ClassHierarchy` as it's over-designed
